Show dashboard associated with habit in habit completion modal bottom sheet
