+++ lib/beamer/locations/dashboards_location.dart
@@ -4,7 +4,6 @@ import 'package:lotti/beamer/beamer_delegates.dart';
 import 'package:lotti/pages/create/complete_habit_dialog.dart';
 import 'package:lotti/pages/create/create_measurement_dialog.dart';
 import 'package:lotti/pages/dashboards/dashboard_page.dart';
-import 'package:lotti/pages/dashboards/dashboards_carousel_page.dart';
 import 'package:lotti/pages/dashboards/dashboards_list_page.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/uuid.dart';
@@ -16,10 +15,7 @@ class DashboardsLocation extends BeamLocation<BeamState> {
   List<String> get pathPatterns => [
         '/dashboards',
         '/dashboards/:dashboardId',
-        '/dashboards/carousel',
-        '/dashboards/carousel/measure/:measurableId',
         '/dashboards/:dashboardId/measure/:measurableId',
-        '/dashboards/carousel/complete_habit/:habitId',
         '/dashboards/:dashboardId/complete_habit/:habitId',
       ];
 
@@ -42,14 +38,7 @@ class DashboardsLocation extends BeamLocation<BeamState> {
           key: ValueKey('dashboards-$dashboardId'),
           child: DashboardPage(dashboardId: dashboardId!),
         ),
-      if (pathContains('carousel'))
-        const BeamPage(
-          key: ValueKey('dashboards-carousel'),
-          child: DashboardCarouselPage(),
-        ),
-      if ((isUuid(dashboardId) || pathContains('carousel')) &&
-          measurableId != null &&
-          isUuid(measurableId))
+      if ((isUuid(dashboardId)) && measurableId != null && isUuid(measurableId))
         BeamPage(
           routeBuilder: (
             BuildContext context,
