+++ lib/pages/create/complete_habit_dialog.dart
@@ -10,7 +10,9 @@ import 'package:lotti/database/logging_db.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/themes/theme.dart';
+import 'package:lotti/utils/platform.dart';
 import 'package:lotti/widgets/charts/utils.dart';
+import 'package:lotti/widgets/dashboards/dashboard_widget.dart';
 import 'package:lotti/widgets/misc/datetime_bottom_sheet.dart';
 import 'package:tinycolor2/tinycolor2.dart';
 import 'package:url_launcher/url_launcher.dart';
@@ -119,115 +121,141 @@ class _HabitDialogState extends State<HabitDialog> {
         if (habitDefinition == null) {
           return const SizedBox.shrink();
         }
+        final timeSpanDays = isDesktop ? 30 : 14;
 
-        return AlertDialog(
-          insetPadding: const EdgeInsets.symmetric(horizontal: 32),
-          contentPadding: const EdgeInsets.only(
-            left: 30,
-            top: 20,
-            right: 10,
-            bottom: 20,
-          ),
-          shape: const RoundedRectangleBorder(
-            borderRadius: BorderRadius.all(Radius.circular(50)),
-          ),
-          backgroundColor: styleConfig().primaryColorLight,
-          actionsAlignment: MainAxisAlignment.spaceBetween,
-          actionsPadding: const EdgeInsets.only(
-            left: 20,
-            right: 20,
-            bottom: 20,
-          ),
-          actions: [
-            TextButton(
-              key: const Key('habit_fail'),
-              onPressed: () => saveHabit(HabitCompletionType.fail),
-              child: Text(
-                localizations.completeHabitFailButton,
-                style: failButtonStyle(),
-              ),
-            ),
-            TextButton(
-              key: const Key('habit_skip'),
-              onPressed: () => saveHabit(HabitCompletionType.skip),
-              child: Text(
-                localizations.completeHabitSkipButton,
-                style: saveButtonStyle()
-                    .copyWith(color: styleConfig().secondaryTextColor),
-              ),
-            ),
-            TextButton(
-              key: const Key('habit_save'),
-              onPressed: () => saveHabit(HabitCompletionType.success),
-              child: Text(
-                localizations.completeHabitSuccessButton,
-                style: saveButtonStyle()
-                    .copyWith(color: styleConfig().primaryColor.darken(25)),
-              ),
-            ),
-          ],
-          content: ConstrainedBox(
-            constraints: const BoxConstraints(maxWidth: 400),
-            child: FormBuilder(
-              key: _formKey,
-              autovalidateMode: AutovalidateMode.onUserInteraction,
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Expanded(
-                        child: Text(
-                          habitDefinition.name,
-                          style: habitCompletionHeaderStyle,
-                        ),
-                      ),
-                      IconButton(
-                        padding: const EdgeInsets.all(10),
-                        icon: const Icon(Icons.close),
-                        onPressed: widget.beamerDelegate.beamBack,
-                      ),
-                    ],
+        final rangeStart = getStartOfDay(
+          DateTime.now().subtract(Duration(days: timeSpanDays)),
+        );
+
+        final rangeEnd = getEndOfToday();
+
+        return SingleChildScrollView(
+          child: Column(
+            children: [
+              const SizedBox(height: 20),
+              AlertDialog(
+                insetPadding: const EdgeInsets.symmetric(horizontal: 32),
+                contentPadding: const EdgeInsets.only(
+                  left: 30,
+                  top: 10,
+                  right: 10,
+                  bottom: 10,
+                ),
+                shape: const RoundedRectangleBorder(
+                  borderRadius: BorderRadius.all(Radius.circular(20)),
+                ),
+                backgroundColor: styleConfig().primaryColorLight,
+                actionsAlignment: MainAxisAlignment.spaceBetween,
+                actionsPadding: const EdgeInsets.only(
+                  left: 20,
+                  right: 20,
+                  bottom: 20,
+                ),
+                actions: [
+                  TextButton(
+                    key: const Key('habit_fail'),
+                    onPressed: () => saveHabit(HabitCompletionType.fail),
+                    child: Text(
+                      localizations.completeHabitFailButton,
+                      style: failButtonStyle(),
+                    ),
                   ),
-                  HabitDescription(habitDefinition),
-                  Padding(
-                    padding: const EdgeInsets.only(right: 20),
+                  TextButton(
+                    key: const Key('habit_skip'),
+                    onPressed: () => saveHabit(HabitCompletionType.skip),
+                    child: Text(
+                      localizations.completeHabitSkipButton,
+                      style: saveButtonStyle()
+                          .copyWith(color: styleConfig().secondaryTextColor),
+                    ),
+                  ),
+                  TextButton(
+                    key: const Key('habit_save'),
+                    onPressed: () => saveHabit(HabitCompletionType.success),
+                    child: Text(
+                      localizations.completeHabitSuccessButton,
+                      style: saveButtonStyle().copyWith(
+                        color: styleConfig().primaryColor.darken(25),
+                      ),
+                    ),
+                  ),
+                ],
+                content: ConstrainedBox(
+                  constraints: const BoxConstraints(
+                    maxWidth: 600,
+                    minWidth: 280,
+                  ),
+                  child: FormBuilder(
+                    key: _formKey,
+                    autovalidateMode: AutovalidateMode.onUserInteraction,
                     child: Column(
+                      mainAxisSize: MainAxisSize.min,
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
-                        inputSpacer,
-                        DateTimeField(
-                          dateTime: _started,
-                          labelText: localizations.addHabitDateLabel,
-                          style: newInputStyle().copyWith(color: Colors.black),
-                          setDateTime: (picked) {
-                            setState(() {
-                              _startReset = true;
-                              _started = picked;
-                            });
-                          },
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                          children: [
+                            Expanded(
+                              child: Text(
+                                habitDefinition.name,
+                                style: habitCompletionHeaderStyle,
+                              ),
+                            ),
+                            IconButton(
+                              padding: const EdgeInsets.all(10),
+                              icon: const Icon(Icons.close),
+                              onPressed: widget.beamerDelegate.beamBack,
+                            ),
+                          ],
                         ),
-                        inputSpacer,
-                        FormBuilderTextField(
-                          initialValue: '',
-                          key: const Key('habit_comment_field'),
-                          decoration: createDialogInputDecoration(
-                            labelText: localizations.addHabitCommentLabel,
+                        HabitDescription(habitDefinition),
+                        Padding(
+                          padding: const EdgeInsets.only(right: 20),
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              inputSpacer,
+                              DateTimeField(
+                                dateTime: _started,
+                                labelText: localizations.addHabitDateLabel,
+                                style: newInputStyle()
+                                    .copyWith(color: Colors.black),
+                                setDateTime: (picked) {
+                                  setState(() {
+                                    _startReset = true;
+                                    _started = picked;
+                                  });
+                                },
+                              ),
+                              inputSpacer,
+                              FormBuilderTextField(
+                                initialValue: '',
+                                key: const Key('habit_comment_field'),
+                                decoration: createDialogInputDecoration(
+                                  labelText: localizations.addHabitCommentLabel,
+                                ),
+                                minLines: 1,
+                                maxLines: 10,
+                                keyboardAppearance: keyboardAppearance(),
+                                style: newInputStyle()
+                                    .copyWith(color: Colors.black),
+                                name: 'comment',
+                              ),
+                            ],
                           ),
-                          minLines: 1,
-                          maxLines: 10,
-                          keyboardAppearance: keyboardAppearance(),
-                          style: newInputStyle().copyWith(color: Colors.black),
-                          name: 'comment',
                         ),
                       ],
                     ),
                   ),
-                ],
+                ),
               ),
-            ),
+              if (habitDefinition.dashboardId != null)
+                DashboardWidget(
+                  rangeStart: rangeStart,
+                  rangeEnd: rangeEnd,
+                  dashboardId: habitDefinition.dashboardId!,
+                ),
+            ],
           ),
         );
       },
