+++ lib/pages/settings/habits/habit_details_page.dart
@@ -15,11 +15,10 @@ import 'package:lotti/pages/settings/form_text_field.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/app_bar/title_app_bar.dart';
 import 'package:lotti/widgets/habits/habit_category.dart';
+import 'package:lotti/widgets/habits/habit_dashboard.dart';
 import 'package:lotti/widgets/misc/datetime_bottom_sheet.dart';
 import 'package:lotti/widgets/settings/form/form_switch.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
-
-import '../../../widgets/habits/habit_dashboard.dart';
 
 class HabitDetailsPage extends StatelessWidget {
   const HabitDetailsPage({super.key});
